+++ system/admin/bootstrap.php
@@ -1,5 +1,5 @@
 <?php declare(strict_types=1);
 
-\define('DASHBOARD_INDEX_KEY', '');
-
+require_once ADMIN_SYS_ROOT.DS.'functions.php';
+require_once ADMIN_SYS_ROOT.DS.'category.php';
 require_once ADMIN_SYS_ROOT.DS.'dispatcher.php';
